Introduces a database data type mapper where the Json Data Types are mapped to the data types of the specific database. For this to happen, the DBUnitConfiguration is used to get the DataTypeFactory (database specific).

+++ impl/src/main/java/org/jboss/arquillian/persistence/dbunit/dataset/yaml/YamlDataSetProducer.java
@@ -179,7 +179,7 @@ public class YamlDataSetProducer implements IDataSetProducer
          DataType dataType = null;
 
          if (value instanceof Integer) {
-            dataType = DataType.INTEGER;
+            dataType = DataType.BIGINT;
          } else {
             dataType = DataType.VARCHAR;
          }
